refactor: move SaveCommand to its own module
