+++ src/services/statefuls/persist/mod.rs
@@ -1,5 +1,6 @@
 pub mod endec;
 pub mod actor;
+pub mod save_command;
 
 use std::collections::HashMap;
