+++ src/services/statefuls/persist/save_command.rs
@@ -1,0 +1,10 @@
+use crate::services::statefuls::cache::CacheEntry;
+
+pub enum SaveCommand {
+    LocalShardSize {
+        table_size: usize,
+        expiry_size: usize,
+    },
+    SaveChunk(Vec<CacheEntry>),
+    StopSentinel,
+}
